Make --running wildcard check to expect only one container to be up
In case of Swarm services, it is not very useful, if passing wildcard as
the container name, requires all matching containers to run.

The Swarm containers are named using '[stack]_[service].id' schema, so
checking '[stack]_[service].*' wildcard can be useful for verifying that
a Swarm service is running on the given host.

However, this change of logic is necessary here to avoid old instances
of the service causing the check to fail.